minor(judger): move task.py into subfolder
